Move collection manager to another module
