+++ test/conftest.py
@@ -5,8 +5,10 @@ import pytest
 
 @pytest.fixture(autouse=True)
 def do_not_run_commands():
-    with mock.patch('ansible_builder.main.run_command', return_value=True) as cmd_mock:
-        yield cmd_mock
+    cmd_mock = mock.MagicMock(return_value=True)
+    with mock.patch('ansible_builder.main.run_command', new=cmd_mock):
+        with mock.patch('ansible_builder.collections.run_command', new=cmd_mock):
+            yield cmd_mock
 
 
 @pytest.fixture
